// src/models/history/HistoryResponse.ts
export interface HistoryResponse {
    id: number
    file_name: string
    row_count: number
    upload_time: string,
    file_type_name: string
    uid: string
    user_id: number
    fullname: string
    data_type_id: number
    account_type_id: number
    file_size: number
    status: boolean
    created_at: string
    inserted_count: number
    skipped_count: number
}